Delete /documents/ in UI, it’s all entities now.
