+++ ui/src/screens/EntityScreen/EntityScreen.jsx
@@ -1,27 +1,59 @@
 import React, { Component } from 'react';
-import { Redirect, withRouter } from 'react-router';
+import { withRouter } from 'react-router';
+import { defineMessages, injectIntl } from 'react-intl';
 import queryString from 'query-string';
 
 import Screen from 'src/components/Screen/Screen';
 import EntityContextLoader from 'src/components/Entity/EntityContextLoader';
 import { compose } from 'redux';
 import { connect } from 'react-redux';
-// import EntityToolbar from 'src/components/Entity/EntityToolbar';
+import EntityToolbar from 'src/components/Entity/EntityToolbar';
 import EntityHeading from 'src/components/Entity/EntityHeading';
 import EntityInfoMode from 'src/components/Entity/EntityInfoMode';
 import EntityViews from 'src/components/Entity/EntityViews';
 import LoadingScreen from 'src/components/Screen/LoadingScreen';
 import ErrorScreen from 'src/components/Screen/ErrorScreen';
-import { DualPane, Breadcrumbs } from 'src/components/common';
+import { DualPane, Breadcrumbs, SearchBox } from 'src/components/common';
+import Query from 'src/app/Query';
 import {
   selectEntity, selectEntityReference, selectEntityView,
 } from 'src/selectors';
 
 
+const messages = defineMessages({
+  placeholder: {
+    id: 'documents.screen.filter',
+    defaultMessage: 'Search in {label}',
+  },
+});
+
+
 class EntityScreen extends Component {
+  static SEARCHABLES = ['Pages', 'Table', 'Folder', 'Package', 'Workbook'];
+
+  constructor(props) {
+    super(props);
+    this.onSearch = this.onSearch.bind(this);
+  }
+
+  onSearch(queryText) {
+    const { history, location, query } = this.props;
+    const parsedHash = queryString.parse(location.hash);
+    const newQuery = query.setString('q', queryText);
+    parsedHash['preview:id'] = undefined;
+    parsedHash['preview:type'] = undefined;
+    parsedHash['preview:mode'] = undefined;
+    parsedHash.page = undefined;
+    history.push({
+      pathname: location.pathname,
+      search: newQuery.toLocation(),
+      hash: queryString.stringify(parsedHash),
+    });
+  }
+
   render() {
     const {
-      entity, entityId, activeMode, location,
+      entity, entityId, activeMode, intl, query,
     } = this.props;
     if (entity.isError) {
       return <ErrorScreen error={entity.error} />;
@@ -33,24 +65,30 @@ class EntityScreen extends Component {
         </EntityContextLoader>
       );
     }
-    if (entity.schema.isDocument()) {
-      return (
-        <Redirect to={{ ...location, pathname: `/documents/${entityId}` }} />
-      );
-    }
+
+    const title = entity.getFirst('title') || entity.getFirst('fileName') || entity.getCaption();
+    const hasSearch = entity.schema.isAny(EntityScreen.SEARCHABLES);
+    const operation = !hasSearch ? undefined : (
+      <SearchBox
+        onSearch={this.onSearch}
+        searchPlaceholder={intl.formatMessage(messages.placeholder, { label: title })}
+        searchText={query.getString('q')}
+      />
+    );
+
     const breadcrumbs = (
-      <Breadcrumbs>
+      <Breadcrumbs operation={operation}>
         <Breadcrumbs.Collection collection={entity.collection} />
         <Breadcrumbs.Entity entity={entity} />
       </Breadcrumbs>
     );
-    const description = entity.getProperty('description').toString(' ');
     return (
       <EntityContextLoader entityId={entityId}>
-        <Screen title={entity.name} description={description}>
+        <Screen title={title}>
           {breadcrumbs}
           <DualPane>
             <DualPane.InfoPane className="with-heading">
+              <EntityToolbar entity={entity} isPreview={false} />
               <EntityHeading entity={entity} isPreview={false} />
               <div className="pane-content">
                 <EntityInfoMode entity={entity} isPreview={false} />
@@ -79,6 +117,7 @@ const mapStateToProps = (state, ownProps) => {
     entityId,
     reference,
     entity: selectEntity(state, entityId),
+    query: Query.fromLocation('entities', location, {}, 'document'),
     activeMode: selectEntityView(state, entityId, hashQuery.mode, false),
   };
 };
@@ -86,4 +125,5 @@ const mapStateToProps = (state, ownProps) => {
 export default compose(
   withRouter,
   connect(mapStateToProps),
+  injectIntl,
 )(EntityScreen);
