+++ ui/src/util/getEntityLink.js
@@ -1,7 +1,3 @@
 export default function getEntityLink(entity) {
-  if (entity.id && entity.schema) {
-    const resource = entity.schema.isDocument() ? 'documents' : 'entities';
-    return `/${resource}/${entity.id}`;
-  }
-  return null;
+  return entity.id ? `/entities/${entity.id}`: null;
 }
